project-maker-bot/requirements.txt
Flask==3.0.3
pydantic==2.9.2
email-validator==2.2.0
requests==2.32.3
PyGithub==2.4.0
python-dotenv==1.0.1
